refactor: move handler functions into handlers.h/.cpp
